+++ cryptol-remote-api/CHANGELOG.md
@@ -1,6 +1,14 @@
 # Revision history for `cryptol-remote-api` and `cryptol-eval-server`
 
-## 2.13.0 -- YYYY-MM-DD
+## 2.13.2 -- YYYY-MM-DD
+
+* Add more fields (such as `pragmas`, `parameter`, `module`, and `infix`) to
+  the response to the RPC `visible names` method.
+* Do not error if `visible names` is called when a parameterized module is
+  loaded (this used to cause the appearance of the server hanging in such a
+  case).
+
+## 2.13.0 -- 2022-05-17
 
 * v2.13.0 release in tandem with the Cryptol 2.13.0 release. See the Cryptol
   2.13.0 release notes for relevant Cryptol changes. No notable changes to the